app/requirements.txt
fastapi==0.95.1
mangum==0.17.0
uvicorn==0.22.0
python-dotenv==1.0.0
boto3==1.28.38
pytest==7.3.1
pytest-cov==4.1.0
httpx==0.24.1
requests==2.31.0